+++ axelor-account/src/main/java/com/axelor/apps/account/db/repo/AccountingReportManagementRepository.java
@@ -21,6 +21,8 @@ import javax.persistence.PersistenceException;
 
 import com.axelor.apps.account.db.AccountingReport;
 import com.axelor.apps.account.service.AccountingReportService;
+import com.axelor.db.JPA;
+import com.axelor.exception.service.TraceBackService;
 import com.google.inject.Inject;
 
 public class AccountingReportManagementRepository extends AccountingReportRepository {
@@ -40,6 +42,9 @@ public class AccountingReportManagementRepository extends AccountingReportReposi
 			
 			return super.save(accountingReport);
 		} catch (Exception e) {
+            JPA.em().getTransaction().commit();
+            JPA.runInTransaction(() -> TraceBackService.trace(e));
+            JPA.em().getTransaction().begin();
 			throw new PersistenceException(e.getLocalizedMessage());
 		}
 	}